Capturing Enemy Towers.
Massive Milestone and core feature pillar of game.
When enemy towers are at low hp they can now be caught by dropping a poke ball on them ... run the check to see if a slot is available and if one is free, use the tower info struct + pokedex data table to create a fresh "party version" tower - then destroy the enemy tower.